requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.2.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
kiwisolver==1.4.4
magic-filter==1.0.9
matplotlib==3.7.1
mouse==0.7.1
multidict==6.0.4
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
six==1.16.0
sniffio==1.3.0
typing_extensions==4.5.0
yarl==1.9.2
